Foundations of Project Management
This course is the first in a series of six to equip you with the skills you need to apply to introductory-level roles in project management. Project managers play a key role in leading, planning and implementing critical projects to help their organizations succeed. In this course, you’ll discover foundational project management terminology and gain a deeper understanding of  the role and responsibilities of a project manager. We’ll also introduce you to the kinds of jobs you might pursue after completing this program. Throughout the program, you’ll learn from current Google project managers, who can provide you with a multi-dimensional educational experience that will help you build your skills  for on-the-job application. 

Learners who complete this program should be equipped to apply for introductory-level jobs as project managers. No previous experience is necessary.

By the end of this course, you will be able to:
- Define project management and describe what constitutes a project.
- Explore project management roles and responsibilities across a variety of industries.
- Detail the core skills that help a project manager be successful.
- Describe the life cycle of a project and explain the significance of each phase.
- Compare different program management methodologies and approaches and determine which is most effective for a given project.
- Define organizational structure and culture and explain how it impacts project management. 
- Define change management and describe the role of the project manager in the process.